refactor: initial date completion tests
