+++ crates/lsp/tests/completion/mod.rs
@@ -1,0 +1,1 @@
+mod date_completion;
