+++ crates/lsp/tests/main.rs
@@ -1,0 +1,1 @@
+mod completion;
